Add Google Analytics and a simple privacy policy
This commit includes a crude privacy policy to comply (and for the sake of
honesty) with Google's terms and conditions. Also, the image comment.png is
removed as it was not longer used and also was infringing it's license by not
providing credit to it's source.

All in all, this commit partially adresses #9.