Start major refactor
This commit starts a major refactor of the source code. In addition, it
implements some basic routing to build upon, starting the path towards
addressing #8.

While this commit compiles, it is a regression in terms of functionality (no
landing page, no login button, no topbar, etc). However, the foundation on top
of which it is possible to build the previous functionaliy is provided here.

More specifically, the following changes are implemented:

* Separation of Model, Update and View onto their own files for the App
component. These files are now under the App domain (App.Model, etc).

* The interaction with the Reddit API is separated into the module RedditAPI.
Originally this module was intended to keep it's own state and handle all the
intricancies of the requests (can't make requests without token, can't request
saved posts without username, etc), so that the Main / App components don't
really need to know about the underlying API to comunicate with Reddit. However,
in this particular commit, while there's some state in RedditAPI.Session that is
handled by RedditAPI.Update, the flow in the app is somewhat contrived: A change
in the location can potentially lead to acquiring a token (coming back from
reddit), which is handled by a Message routed to RedditAPI.Update; acquiring the
token inmediatly generates a Cmd to get the username within the mentioned
update; when the response from this request arrives it is in the form of a
Message directed to RedditAPI.Update, and so it is wrapped in APIMsg; however,
the message is intercepted in Main.elm and if it's a username response, a new
Cmd to ask for the saved posts is generated. In summary, the code in Main asks
for the saved posts as soon as the response with the username directed at the
API state arrives. Ideally, Main should only be concerned about creating a
command to retrieve the saved posts as soon as a token is acquired. The
RedditAPI module should handle making the appropiate requests and finally
generating some Message with the saved posts for the Main module. Moreoever, in
this way RedditAPI can also keep track of the number of requests made so that
the limits imposed by reddit are not exceeded.

* The framework for routing is provided. Right now we can track the current
"page" as in "Landing Page", "About", "App", etc, but the view isn't affected.
We also track the page number of the list of items in the app.