Implements basic Landing Page
A simple landing page describing the app is created. Right now it's rendered in
Elm, maybe it could be moved to a static html file? This commit closes #10.

Also, further use of elm-mdl is made, in relation to issue #1.

For the time being, if we continue to use elm-mdl, this closes #2 as the top bar
as it used to be doesn't exist anymore.